+++ packages/asset-swapper/contracts/src/ERC20BridgeSampler.sol
@@ -34,6 +34,7 @@ import "./MStableSampler.sol";
 import "./MooniswapSampler.sol";
 import "./NativeOrderSampler.sol";
 import "./ShellSampler.sol";
+import "./SmoothySampler.sol";
 import "./TwoHopSampler.sol";
 import "./UniswapSampler.sol";
 import "./UniswapV2Sampler.sol";
@@ -55,6 +56,7 @@ contract ERC20BridgeSampler is
     MultiBridgeSampler,
     NativeOrderSampler,
     ShellSampler,
+    SmoothySampler,
     TwoHopSampler,
     UniswapSampler,
     UniswapV2Sampler,