+++ packages/asset-swapper/src/utils/market_operation_utils/orders.ts
@@ -89,6 +89,8 @@ export function getErc20BridgeSourceToBridgeSource(source: ERC20BridgeSource): s
             return encodeBridgeSourceId(BridgeProtocol.Dodo, 'Dodo');
         case ERC20BridgeSource.Kyber:
             return encodeBridgeSourceId(BridgeProtocol.Kyber, 'Kyber');
+        case ERC20BridgeSource.KyberDmm:
+            return encodeBridgeSourceId(BridgeProtocol.UniswapV2, 'KyberDmm');
         case ERC20BridgeSource.LiquidityProvider:
             // "LiquidityProvider" is too long to encode (17 characters).
             return encodeBridgeSourceId(BridgeProtocol.Unknown, 'LP');
@@ -126,6 +128,12 @@ export function getErc20BridgeSourceToBridgeSource(source: ERC20BridgeSource): s
             return encodeBridgeSourceId(BridgeProtocol.Curve, 'Belt');
         case ERC20BridgeSource.Ellipsis:
             return encodeBridgeSourceId(BridgeProtocol.Curve, 'Ellipsis');
+        case ERC20BridgeSource.Component:
+            return encodeBridgeSourceId(BridgeProtocol.Shell, 'Component');
+        case ERC20BridgeSource.Smoothy:
+            return encodeBridgeSourceId(BridgeProtocol.Curve, 'Smoothy');
+        case ERC20BridgeSource.Saddle:
+            return encodeBridgeSourceId(BridgeProtocol.Nerve, 'Saddle');
         default:
             throw new Error(AggregationError.NoBridgeForSource);
     }
@@ -153,6 +161,8 @@ export function createBridgeDataForBridgeOrder(order: OptimizedMarketBridgeOrder
         case ERC20BridgeSource.Nerve:
         case ERC20BridgeSource.Belt:
         case ERC20BridgeSource.Ellipsis:
+        case ERC20BridgeSource.Smoothy:
+        case ERC20BridgeSource.Saddle:
             const curveFillData = (order as OptimizedMarketBridgeOrder<CurveFillData>).fillData;
             bridgeData = encoder.encode([
                 curveFillData.pool.poolAddress,
@@ -176,6 +186,7 @@ export function createBridgeDataForBridgeOrder(order: OptimizedMarketBridgeOrder
         case ERC20BridgeSource.Linkswap:
         case ERC20BridgeSource.PancakeSwap:
         case ERC20BridgeSource.BakerySwap:
+        case ERC20BridgeSource.KyberDmm:
             const uniswapV2FillData = (order as OptimizedMarketBridgeOrder<UniswapV2FillData>).fillData;
             bridgeData = encoder.encode([uniswapV2FillData.router, uniswapV2FillData.tokenAddressPath]);
             break;
@@ -200,6 +211,7 @@ export function createBridgeDataForBridgeOrder(order: OptimizedMarketBridgeOrder
             bridgeData = encoder.encode([dodoV2FillData.poolAddress, dodoV2FillData.isSellBase]);
             break;
         case ERC20BridgeSource.Shell:
+        case ERC20BridgeSource.Component:
             const shellFillData = (order as OptimizedMarketBridgeOrder<ShellFillData>).fillData;
             bridgeData = encoder.encode([shellFillData.poolAddress]);
             break;
@@ -299,14 +311,18 @@ export const BRIDGE_ENCODERS: {
     [ERC20BridgeSource.Nerve]: curveEncoder,
     [ERC20BridgeSource.Belt]: curveEncoder,
     [ERC20BridgeSource.Ellipsis]: curveEncoder,
+    [ERC20BridgeSource.Smoothy]: curveEncoder,
+    [ERC20BridgeSource.Saddle]: curveEncoder,
     // UniswapV2 like, (router, address[])
     [ERC20BridgeSource.Bancor]: routerAddressPathEncoder,
     [ERC20BridgeSource.UniswapV2]: routerAddressPathEncoder,
     [ERC20BridgeSource.SushiSwap]: routerAddressPathEncoder,
     [ERC20BridgeSource.CryptoCom]: routerAddressPathEncoder,
     [ERC20BridgeSource.Linkswap]: routerAddressPathEncoder,
+    [ERC20BridgeSource.KyberDmm]: routerAddressPathEncoder,
     // Generic pools
     [ERC20BridgeSource.Shell]: poolEncoder,
+    [ERC20BridgeSource.Component]: poolEncoder,
     [ERC20BridgeSource.Mooniswap]: poolEncoder,
     [ERC20BridgeSource.Eth2Dai]: poolEncoder,
     [ERC20BridgeSource.MStable]: poolEncoder,